[OfllineStorage]
1) edit the color background in details screen using UIImageColors cocoaPods
2) work on speed up the displaying the first view
3) save 20 photo in the  offline storage

+++ IntraZeroTask/Model/Photo.swift
@@ -11,8 +11,8 @@ struct Photo: Codable , Hashable ,Identifiable  {
 
     let id: String
     let author: String
-    let width: Int
-    let height: Int
+    let width: Int16
+    let height: Int16
     let url: String
     let downloadUrl: String
